Move users table to separate file
