+++ src/users/users.repository.ts
@@ -1,10 +1,10 @@
 // users.repository.ts
 // in charge of database interactions
 
-import { users } from './users.schema';
+import { eq } from 'drizzle-orm';
 import { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import { UserToCreate } from '@/users/users.schema';
-import { eq } from 'drizzle-orm';
+import { users } from '@/users/users.model';
 
 export class UsersRepository {
   constructor(private readonly db: PostgresJsDatabase) {}
