+++ src/users/users.schema.ts
@@ -1,18 +1,6 @@
 import { Type } from '@sinclair/typebox';
-import { sql } from 'drizzle-orm';
-import { date, pgTable, serial, text } from 'drizzle-orm/pg-core';
 import { createInsertSchema, createSelectSchema } from 'drizzle-typebox';
-
-export const users = pgTable('users', {
-  id: serial('id').primaryKey(),
-  email: text('email').unique().notNull(),
-  bio: text('bio'),
-  image: text('image'),
-  password: text('password').notNull(),
-  username: text('username').notNull(),
-  created_at: date('created_at').default(sql`CURRENT_DATE`),
-  updated_at: date('updated_at').default(sql`CURRENT_DATE`),
-});
+import { users } from '@users/users.model';
 
 // Schema for inserting a user - can be used to validate API requests
 const insertUserSchemaRaw = createInsertSchema(users);
